+++ CHANGELOG.md
@@ -9,7 +9,7 @@
 - If file is loaded into edit mode with face selected, that face is used to set active image
 - Default material is always scale 1 (initial texturing makes more sense)
 - Fix grid hotkeys not working in object mode when file first loaded (they worked once toggled in and out of edit mode)
-- Face UV Mode (Press T to snap texture to edge based on mouse position)
+- Face UV Mode (Press T to snap texture to edge based on mouse position) (removed UV Rotation panel)
 - Fix UV shortcut buttons
 
 ## 1.0.2